Fixed SendingVerpasst To Load Content Rows Properly
Before it only loaded the content into one row and had
a bunch of dummy rows. It now populates the rows properly
and displays them the same way as they are presented on the web.
It also updates the thumbnail image size to work for landscape. All
images are only available for landscape.